add:res discount

+++ src/main/java/com/example/udpm14sellcomputerpartsbackend/model/dto/ProductDto.java
@@ -30,7 +30,7 @@ public class ProductDto {
 
     private LocalDateTime createDate;
     private LocalDateTime updateDate;
-    private Float discount;
+    private Long discount;
     private String description;
     private StatusEnum status;
     private Long categoryId;
